Not for merging - test code for forks
For a repo that's a fork, we could take the subsection of contributors that have committed on the fork but not parent as that repos contributors

But I do not believe this is valid, because the contributors to a fork should include all the people who helped build the parent. Also, github fork workflows etcetc. Also, given current methods, it's nigh impossible to determine if someone has contributed to a fork if they also happen to have an upstream comment, as they'd cancel themselves out